django/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.57.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
MarkupSafe==2.1.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
